+++ packages/dappmanager/src/params.ts
@@ -136,7 +136,7 @@ const params = {
   CHECK_DISK_USAGE_DAEMON_INTERVAL: 1 * MINUTE,
   NAT_RENEWAL_DAEMON_INTERVAL: 1 * HOUR,
   NSUPDATE_DAEMON_INTERVAL: 1 * HOUR,
-  ETH_METRICS_DAEMON_INTERVAL: 1 * HOUR,
+  ETH_METRICS_DAEMON_INTERVAL: 50 * MINUTE,
 
   // IPFS parameters
   IPFS_HOST: process.env.IPFS_HOST || process.env.IPFS_REDIRECT,
